#181 Change mint module to default cosmos module
