+++ app/fee.go
@@ -1,6 +1,9 @@
 package app
 
-import sdk "github.com/cosmos/cosmos-sdk/types"
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"log"
+)
 
 var noCoins = sdk.Coins{}
 
@@ -8,7 +11,8 @@ var noCoins = sdk.Coins{}
 // fee collection keeper used only for testing
 type NoopFeeCollectionKeeper struct{}
 
-func (fck NoopFeeCollectionKeeper) AddCollectedFees(sdk.Context, sdk.Coins) sdk.Coins {
+func (fck NoopFeeCollectionKeeper) AddCollectedFees(_ sdk.Context, c sdk.Coins) sdk.Coins {
+	log.Println(c.String())
 	return noCoins
 }
