+++ app/genesis.go
@@ -14,11 +14,6 @@ import (
 	"github.com/cybercongress/cyberd/types/coin"
 	"github.com/pkg/errors"
 	tmtypes "github.com/tendermint/tendermint/types"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"sort"
-	"strings"
 	"time"
 )
 
@@ -91,12 +86,37 @@ const (
 	defaultUnbondingTime = 60 * 60 * 24 * 3 * time.Second
 )
 
-// DefaultParams returns a default set of parameters.
-func DefaultStakingParams() types.Params {
-	return types.Params{
-		UnbondingTime: defaultUnbondingTime,
-		MaxValidators: 146,
-		BondDenom:     coin.CBD,
+// todo set params for each module
+// NewDefaultGenesisState generates the default state for cyberd.
+func NewDefaultGenesisState() GenesisState {
+	return GenesisState{
+		Accounts: nil,
+		AuthData: auth.GenesisState{
+			Params: auth.Params{
+				MaxMemoCharacters: 256,
+			},
+		},
+		MintData: mint.GenesisState{
+			Params: mint.Params{
+				MintDenom:           coin.CBD,
+				InflationRateChange: sdk.NewDecWithPrec(0, 2),
+				InflationMax:        sdk.NewDecWithPrec(200, 2),
+				InflationMin:        sdk.NewDecWithPrec(200, 2),
+				GoalBonded:          sdk.NewDecWithPrec(99, 2),
+				BlocksPerYear:       uint64(60 * 60 * 24 * 365), // assuming 1 second block times
+			},
+		},
+		StakingData: staking.GenesisState{
+			Pool: staking.InitialPool(),
+			Params: types.Params{
+				UnbondingTime: defaultUnbondingTime,
+				MaxValidators: 146,
+				BondDenom:     coin.CBD,
+			},
+		},
+		SlashingData: slashing.DefaultGenesisState(),
+		DistrData:    distr.DefaultGenesisState(),
+		GenTxs:       nil,
 	}
 }
 
@@ -139,19 +159,15 @@ func CyberdAppGenState(cdc *codec.Codec, genDoc tmtypes.GenesisDoc, appGenTxs []
 	return genesisState, nil
 }
 
-// todo set params for each module
-// NewDefaultGenesisState generates the default state for cyberd.
-func NewDefaultGenesisState() GenesisState {
-	return GenesisState{
-		Accounts: nil,
-		StakingData: staking.GenesisState{
-			Pool:   staking.InitialPool(),
-			Params: DefaultStakingParams(),
-		},
-		SlashingData: slashing.DefaultGenesisState(),
-		DistrData:    distr.DefaultGenesisState(),
-		GenTxs:       nil,
+// CyberdAppGenState but with JSON
+func CyberdAppGenStateJSON(cdc *codec.Codec, genDoc tmtypes.GenesisDoc, appGenTxs []json.RawMessage) (
+	appState json.RawMessage, err error) {
+	// create the final app state
+	genesisState, err := CyberdAppGenState(cdc, genDoc, appGenTxs)
+	if err != nil {
+		return nil, err
 	}
+	return codec.MarshalJSONIndent(cdc, genesisState)
 }
 
 // CyberdValidateGenesisState ensures that the genesis state obeys the expected invariants
@@ -187,116 +203,4 @@ func validateGenesisStateAccounts(accs []GenesisAccount) (err error) {
 		addrMap[strAddr] = true
 	}
 	return
-}
-
-// CyberdAppGenState but with JSON
-func CyberdAppGenStateJSON(cdc *codec.Codec, genDoc tmtypes.GenesisDoc, appGenTxs []json.RawMessage) (
-	appState json.RawMessage, err error) {
-	// create the final app state
-	genesisState, err := CyberdAppGenState(cdc, genDoc, appGenTxs)
-	if err != nil {
-		return nil, err
-	}
-	return codec.MarshalJSONIndent(cdc, genesisState)
-}
-
-// todo move to other place??
-// CollectStdTxs processes and validates application's genesis StdTxs and returns
-// the list of appGenTxs, and persistent peers required to generate genesis.json.
-func CollectStdTxs(cdc *codec.Codec, moniker string, genTxsDir string, genDoc tmtypes.GenesisDoc) (
-	appGenTxs []auth.StdTx, persistentPeers string, err error) {
-
-	var fos []os.FileInfo
-	fos, err = ioutil.ReadDir(genTxsDir)
-	if err != nil {
-		return appGenTxs, persistentPeers, err
-	}
-
-	// prepare a map of all accounts in genesis state to then validate
-	// against the validators addresses
-	var appState GenesisState
-	if err := cdc.UnmarshalJSON(genDoc.AppState, &appState); err != nil {
-		return appGenTxs, persistentPeers, err
-	}
-	addrMap := make(map[string]GenesisAccount, len(appState.Accounts))
-	for i := 0; i < len(appState.Accounts); i++ {
-		acc := appState.Accounts[i]
-		strAddr := string(acc.Address)
-		addrMap[strAddr] = acc
-	}
-
-	// addresses and IPs (and port) validator server info
-	var addressesIPs []string
-
-	for _, fo := range fos {
-		filename := filepath.Join(genTxsDir, fo.Name())
-		if !fo.IsDir() && (filepath.Ext(filename) != ".json") {
-			continue
-		}
-
-		// get the genStdTx
-		var jsonRawTx []byte
-		if jsonRawTx, err = ioutil.ReadFile(filename); err != nil {
-			return appGenTxs, persistentPeers, err
-		}
-		var genStdTx auth.StdTx
-		if err = cdc.UnmarshalJSON(jsonRawTx, &genStdTx); err != nil {
-			return appGenTxs, persistentPeers, err
-		}
-		appGenTxs = append(appGenTxs, genStdTx)
-
-		// the memo flag is used to store
-		// the ip and node-id, for example this may be:
-		// "528fd3df22b31f4969b05652bfe8f0fe921321d5@192.168.2.37:26656"
-		nodeAddrIP := genStdTx.GetMemo()
-		if len(nodeAddrIP) == 0 {
-			return appGenTxs, persistentPeers, fmt.Errorf(
-				"couldn't find node's address and IP in %s", fo.Name())
-		}
-
-		// genesis transactions must be single-message
-		msgs := genStdTx.GetMsgs()
-		if len(msgs) != 1 {
-
-			return appGenTxs, persistentPeers, errors.New(
-				"each genesis transaction must provide a single genesis message")
-		}
-
-		msg := msgs[0].(staking.MsgCreateValidator)
-		// validate delegator and validator addresses and funds against the accounts in the state
-		delAddr := msg.DelegatorAddr.String()
-		valAddr := sdk.AccAddress(msg.ValidatorAddr).String()
-
-		delAcc, delOk := addrMap[delAddr]
-		_, valOk := addrMap[valAddr]
-
-		var accsNotInGenesis []string
-		if !delOk {
-			accsNotInGenesis = append(accsNotInGenesis, delAddr)
-		}
-		if !valOk {
-			accsNotInGenesis = append(accsNotInGenesis, valAddr)
-		}
-		if len(accsNotInGenesis) != 0 {
-			return appGenTxs, persistentPeers, fmt.Errorf(
-				"account(s) %v not in genesis.json: %+v", strings.Join(accsNotInGenesis, " "), addrMap)
-		}
-
-		if sdk.NewInt(delAcc.Amount).LT(msg.Value.Amount) {
-			return appGenTxs, persistentPeers, fmt.Errorf(
-				"insufficient fund for delegation %v: %v < %v",
-				delAcc.Address, sdk.NewInt(delAcc.Amount), msg.Value.Amount,
-			)
-		}
-
-		// exclude itself from persistent peers
-		if msg.Description.Moniker != moniker {
-			addressesIPs = append(addressesIPs, nodeAddrIP)
-		}
-	}
-
-	sort.Strings(addressesIPs)
-	persistentPeers = strings.Join(addressesIPs, ",")
-
-	return appGenTxs, persistentPeers, nil
 }
